Castor3D: First implementation for pass components.

+++ include/Core/Castor3D/Shader/Ubos/SceneUbo.hpp
@@ -41,7 +41,7 @@ namespace castor3d
 				, sdw::Float const & depth
 				, sdw::Vec3 const & colour
 				, sdw::Float const & alpha
-				, sdw::Float const & accumulationOperator )const;
+				, sdw::UInt const & accumulationOperator )const;
 
 		private:
 			using sdw::StructInstance::getMember;
